mid debugging
