+++ packages/backend/src/modules/message/message.module.ts
@@ -1,21 +1,5 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { DbModule } from "../db/db.module";
-import { OpportunityModule } from "../opportunity/opportunity.module";
-import { EmailService } from "./email.service";
-import { MessageController } from "./message.controller";
-import { MessageEntity } from "./message.entity";
-import { MessageService } from "./message.service";
-import { MessageThreadEntity } from "./thread.entity";
 
-@Module({
-  imports: [
-    DbModule,
-    TypeOrmModule.forFeature([MessageThreadEntity]),
-    TypeOrmModule.forFeature([MessageEntity]),
-    OpportunityModule,
-  ],
-  providers: [MessageService, EmailService],
-  controllers: [MessageController],
-})
+@Module({})
 export class MessageModule {}
